make access target resolution process more specific
I feel this was originally over-generified. While it is possible to treat the resolution uniformly for field accesses, method calls and constructor calls, it completely hides the fact that we do not need any class hierarchy resolution complexity for constructor calls. Since constructor calls will always be directly to the target owner, i.e. it is impossible to call a super class constructor from the outside, the constructor is always declared within the class itself. I have thus decided to remove the polymorphism from `TargetInfo` and move the resolution logic into `AccessFactory`, where the complex hierarchy resolution now only needs to be applied to fields and methods.
The logic still feels overcomplicated and sub-optimal, e.g. we get all fields/methods of a class as a set and randomly look for a path from `target.desc` to `member`, and if there is EXACTLY ONE match in the path, we accept it. If we would sort the members according to the class hierarchy we would never get to the case that we find a match, but there is another field shadowing / method overriding in the hierarchy between `target` and `member`. However since this part is not completely trivial I decided to save this for a later point in time.
