+++ archunit/src/main/java/com/tngtech/archunit/core/importer/RawAccessRecord.java
@@ -15,32 +15,14 @@
  */
 package com.tngtech.archunit.core.importer;
 
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Objects;
-import java.util.Set;
 
-import com.tngtech.archunit.base.DescribedPredicate;
-import com.tngtech.archunit.base.Optional;
-import com.tngtech.archunit.core.domain.JavaClass;
 import com.tngtech.archunit.core.domain.JavaClassDescriptor;
 import com.tngtech.archunit.core.domain.JavaCodeUnit;
-import com.tngtech.archunit.core.domain.JavaConstructor;
-import com.tngtech.archunit.core.domain.JavaField;
 import com.tngtech.archunit.core.domain.JavaFieldAccess.AccessType;
-import com.tngtech.archunit.core.domain.JavaMethod;
-import com.tngtech.archunit.core.domain.properties.HasDescriptor;
-import com.tngtech.archunit.core.domain.properties.HasName;
-import com.tngtech.archunit.core.domain.properties.HasOwner;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static com.tngtech.archunit.core.domain.JavaConstructor.CONSTRUCTOR_NAME;
-import static com.tngtech.archunit.core.domain.properties.HasName.Predicates.nameMatching;
-import static java.util.Collections.singletonList;
-import static java.util.regex.Pattern.quote;
 
 class RawAccessRecord {
     final CodeUnit caller;
@@ -142,7 +124,7 @@ class RawAccessRecord {
         }
     }
 
-    abstract static class TargetInfo {
+    static class TargetInfo {
         final JavaClassDescriptor owner;
         final String name;
         final String desc;
@@ -151,28 +133,6 @@ class RawAccessRecord {
             this.owner = JavaClassDescriptorImporter.createFromAsmObjectTypeName(owner);
             this.name = name;
             this.desc = desc;
-        }
-
-        <T extends HasName & HasDescriptor & HasOwner<JavaClass>> boolean matches(T member) {
-            if (!name.equals(member.getName()) || !desc.equals(member.getDescriptor())) {
-                return false;
-            }
-            return owner.getFullyQualifiedClassName().equals(member.getOwner().getName()) ||
-                    classHierarchyFrom(member).hasExactlyOneMatchFor(this);
-        }
-
-        private <T extends HasName & HasDescriptor & HasOwner<JavaClass>> ClassHierarchyPath classHierarchyFrom(T member) {
-            return new ClassHierarchyPath(owner, member.getOwner());
-        }
-
-        protected abstract boolean signatureExistsIn(JavaClass javaClass);
-
-        boolean hasMatchingSignatureTo(JavaMethod method) {
-            return method.getName().equals(name) && method.getDescriptor().equals(desc);
-        }
-
-        boolean hasMatchingSignatureTo(JavaConstructor constructor) {
-            return CONSTRUCTOR_NAME.equals(name) && constructor.getDescriptor().equals(desc);
         }
 
         @Override
@@ -197,204 +157,6 @@ class RawAccessRecord {
         @Override
         public String toString() {
             return getClass().getSimpleName() + "{owner='" + owner.getFullyQualifiedClassName() + "', name='" + name + "', desc='" + desc + "'}";
-        }
-
-        private static class ClassHierarchyPath {
-            private final List<JavaClass> path = new ArrayList<>();
-
-            private ClassHierarchyPath(JavaClassDescriptor childType, JavaClass parent) {
-                Set<JavaClass> classesToSearchForChild = parent.getAllSubClasses();
-                Optional<JavaClass> child = tryFind(classesToSearchForChild, nameMatching(quote(childType.getFullyQualifiedClassName())));
-                if (child.isPresent()) {
-                    createPath(child.get(), parent);
-                }
-            }
-
-            private static <T> Optional<T> tryFind(Iterable<T> collection, DescribedPredicate<? super T> predicate) {
-                for (T elem : collection) {
-                    if (predicate.apply(elem)) {
-                        return Optional.of(elem);
-                    }
-                }
-                return Optional.absent();
-            }
-
-            private void createPath(JavaClass child, JavaClass parent) {
-                HierarchyResolutionStrategy hierarchyResolutionStrategy = hierarchyResolutionStrategyFrom(child).to(parent);
-                path.add(child);
-                while (hierarchyResolutionStrategy.hasNext()) {
-                    path.add(hierarchyResolutionStrategy.next());
-                }
-            }
-
-            boolean hasExactlyOneMatchFor(final TargetInfo target) {
-                Set<JavaClass> matching = new HashSet<>();
-                for (JavaClass javaClass : path) {
-                    if (target.signatureExistsIn(javaClass)) {
-                        matching.add(javaClass);
-                    }
-                }
-                return matching.size() == 1;
-            }
-
-            private HierarchyResolutionStrategyCreator hierarchyResolutionStrategyFrom(JavaClass child) {
-                return new HierarchyResolutionStrategyCreator(child);
-            }
-
-            private interface HierarchyResolutionStrategy {
-                boolean hasNext();
-
-                JavaClass next();
-            }
-
-            private static class HierarchyResolutionStrategyCreator {
-                private final JavaClass child;
-
-                private HierarchyResolutionStrategyCreator(JavaClass child) {
-                    this.child = child;
-                }
-
-                public HierarchyResolutionStrategy to(JavaClass parent) {
-                    return parent.isInterface() ?
-                            new InterfaceHierarchyResolutionStrategy(child, parent) :
-                            new ClassHierarchyResolutionStrategy(child, parent);
-                }
-            }
-
-            private static class ClassHierarchyResolutionStrategy implements HierarchyResolutionStrategy {
-                private final JavaClass parent;
-                private JavaClass current;
-
-                private ClassHierarchyResolutionStrategy(JavaClass child, JavaClass parent) {
-                    this.current = child;
-                    this.parent = parent;
-                }
-
-                @Override
-                public boolean hasNext() {
-                    return !current.equals(parent) && current.getSuperClass().isPresent();
-                }
-
-                @Override
-                public JavaClass next() {
-                    current = current.getSuperClass().get();
-                    return current;
-                }
-            }
-
-            private static class InterfaceHierarchyResolutionStrategy implements HierarchyResolutionStrategy {
-                private final Iterator<JavaClass> interfaces;
-                private final JavaClass parent;
-                private JavaClass current;
-
-                private InterfaceHierarchyResolutionStrategy(JavaClass child, JavaClass parent) {
-                    interfaces = interfacesBetween(child, parent);
-                    this.parent = parent;
-                    current = child;
-                }
-
-                private Iterator<JavaClass> interfacesBetween(JavaClass from, JavaClass target) {
-                    Node node = new Node(from);
-                    List<JavaClass> result = new ArrayList<>();
-                    for (Node parent : node.parents) {
-                        result.addAll(parent.to(target));
-                    }
-                    return result.iterator();
-                }
-
-                @Override
-                public boolean hasNext() {
-                    return !current.equals(parent) && interfaces.hasNext();
-                }
-
-                @Override
-                public JavaClass next() {
-                    current = interfaces.next();
-                    return current;
-                }
-            }
-
-            private static class Node {
-                private final JavaClass child;
-                private final Set<Node> parents = new HashSet<>();
-
-                private Node(JavaClass child) {
-                    this.child = child;
-                    for (JavaClass i : child.getInterfaces()) {
-                        parents.add(new Node(i));
-                    }
-                }
-
-                public List<JavaClass> to(JavaClass target) {
-                    if (child.equals(target)) {
-                        return singletonList(child);
-                    }
-                    Set<JavaClass> result = new LinkedHashSet<>();
-                    for (Node parent : parents) {
-                        if (parent.contains(target)) {
-                            result.add(child);
-                            result.addAll(parent.to(target));
-                        }
-                    }
-                    return new ArrayList<>(result);
-                }
-
-                public boolean contains(JavaClass target) {
-                    if (child.equals(target)) {
-                        return true;
-                    }
-                    for (Node parent : parents) {
-                        if (parent.contains(target)) {
-                            return true;
-                        }
-                    }
-                    return false;
-                }
-            }
-        }
-    }
-
-    static class FieldTargetInfo extends TargetInfo {
-        FieldTargetInfo(String owner, String name, String desc) {
-            super(owner, name, desc);
-        }
-
-        @Override
-        protected boolean signatureExistsIn(JavaClass javaClass) {
-            Optional<JavaField> field = javaClass.tryGetField(name);
-            return field.isPresent() && desc.equals(field.get().getDescriptor());
-        }
-    }
-
-    static class ConstructorTargetInfo extends TargetInfo {
-        ConstructorTargetInfo(String owner, String name, String desc) {
-            super(owner, name, desc);
-        }
-
-        @Override
-        protected boolean signatureExistsIn(JavaClass javaClass) {
-            for (JavaConstructor constructor : javaClass.getConstructors()) {
-                if (hasMatchingSignatureTo(constructor)) {
-                    return true;
-                }
-            }
-            return false;
-        }
-    }
-
-    static class MethodTargetInfo extends TargetInfo {
-        MethodTargetInfo(String owner, String name, String desc) {
-            super(owner, name, desc);
-        }
-
-        @Override
-        protected boolean signatureExistsIn(JavaClass javaClass) {
-            for (JavaMethod method : javaClass.getMethods()) {
-                if (hasMatchingSignatureTo(method)) {
-                    return true;
-                }
-            }
-            return false;
         }
     }
